feat: poll add numeric and tag cloud

+++ src/Services/ConferenceManagement/Strive.Core/Services/Poll/Types/Numeric/NumericAnswer.cs
@@ -1,4 +1,4 @@
 ﻿namespace Strive.Core.Services.Poll.Types.Numeric
 {
-    public record NumericAnswer(double Selected) : PollAnswer;
+    public record NumericAnswer(decimal Selected) : PollAnswer;
 }
